Qualification almost works
Made a few changes on that end, only one bug surrounding it still needs to be resolved but besides that it seems to work how i want it to. The bug is: the qualification table doesnt get updated for Q2 en Q3 but new values are added to the table.